Add static folder

+++ convert_db.py
@@ -9,7 +9,8 @@ import sys
 
 MVLENS_FOLDER = "data"
 
-os.remove("database.db")
+if os.path.exists("database.db"):
+    os.remove("database.db")
 db = sqlite3.connect("database.db")
 
 # Create tables
